# opaa-api/src/main/resources/openapi/opaa-api.yaml
openapi: 3.0.3
info:
  title: OPAA API
  description: Open Project AI Assistant REST API
  version: 0.1.0
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
paths:
...
  /api/v1/libraries/{libraryId}/confluence-webhook:
    parameters:
      - name: libraryId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      operationId: receiveConfluenceWebhook
      tags: [indexing]
      summary: Receive a change notification from a Confluence instance or Automation rule
      description: >
        #1140: the only endpoint under /api/v1 a caller reaches without a session - the sender
        has none. The request authenticates itself with the library's webhook secret instead,
        either as Confluence Data Center signs it (header X-Hub-Signature: sha256=<hex HMAC-SHA256
        of the raw body>) or as a Confluence Cloud Automation rule sends it (header
        X-OPAA-Webhook-Secret carrying the secret itself). An unknown library, a library without
        a secret or of another sourceType, and a wrong or missing signature are all answered with
        the same 401; no other information leaves this endpoint. Once authenticated, the page ids
        the body names (page.id, content.id, attachment.pageId, pageId, pageIds) are queued and,
        a few seconds later (opaa.indexing.confluence.webhook.debounce), one WEBHOOK run fetches
        exactly those pages: a page the instance reports as trashed is removed with its
        attachments, a changed page is re-indexed, a 404 or 403 changes nothing (ADR-0023,
        Entscheidung 4 - deletion needs the instance's own positive finding, the notification is
        not one). The event type is deliberately not interpreted. A body naming no page is
        accepted and ignored. Rate-limited per library (opaa.rate-limit.webhook).
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: The sender's own payload; only the page ids named in it are read.
      responses:
        "202":
          description: Accepted and queued (or nothing to queue)
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, status, timestamp]
      properties:
        error:
          type: string
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...